package.json - scripts - pwa - try a different way
I was getting an error, that I'd previously seen when
I was doing the `npx serve` way (a few commits ago)
and earlier, I'd seen that happen with `npx serve`
then go away when I was using `npx http-server`.

Didn't have a good explanation for WHY the difference,
but I wanted to try it this way again, as part of
trial and error for the current error.

Unfortunately the error is still coming up anyways
(and I can't tell why it's happening now, when
it wasn't earlier). Seems to be appearing on master too.

but anyways, this adds a script to do local
testing over HTTPS, using http-server --tls.
(Also a script to generate the cert/key it needs.)
We might want to keep that part, as an easier way
to test the installable PWA over HTTPS on an address
besides localhost