Remove pwa:https script, unsure if useful
For some reason, the service worker was not working when I tested this.
Maybe it has to do with the invalid cert / bad https creds,
and maybe the browser blocked it as a result?

(More specifically it shows the service worker as being received
at 12/31/69, so, something's going wrong there.
When I test the pwa over http - npm run pwa - it works OK, in that case)