Override WorkboxWebpackPlugin the CRACO way
I did a search to see if/how other people are overriding
WorkboxWebpackPlugin configuration in craco.config.js. Seems that the
other error goes away when we do it inside {webpack.configure} instead of
putting it in {plugins}.

However... birdseye is not working as of this commit.
So there are still things to troubleshoot...